refactor: move test files to a __tests__ folder
