+++ src/internal/validation/transaction/__tests__/validate-create-data.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "@jest/globals";
-import { createData } from "./validate-create-data";
 import { CreateTransactionDto } from "@/internal/dto/transaction";
+import { createData } from "../validate-create-data";
 
 describe("transaction create data", () => {
   it("should return false for invalid data", () => {
