+++ src/internal/validation/transaction/__tests__/validate-update-data.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "@jest/globals";
-import { updateData } from "./validate-update-data";
+import { updateData } from "../validate-update-data";
 import { UpdateTransactionDto } from "@/internal/dto/transaction";
 
 describe("transaction update data", () => {
